More refactoring to allow specifying RID to compile
