+++ scripts/linux.py
@@ -8,20 +8,27 @@ from software.freetype import Freetype
 from software.fluidsynth import Fluidsynth
 
 from common import Github, Software, dump_build_notes, ARTIFACT_DIR, Platform, ROOT_DIR
+from common.software import SoftwareImpl
+from common.args import parse_args
+from common.platform import RID_LINUX_X64
 
 import subprocess
 
 if __name__ == "__main__":
-    to_build: list[Software] = [
-        ZStd(),
-        GLFW(),
-        SDL(),
-        OpenAL(),
-        Freetype(),
-        Fluidsynth(),
+    args = parse_args(RID_LINUX_X64)
+
+    to_build: list[SoftwareImpl] = [
+        ZStd,
+        GLFW,
+        SDL,
+        OpenAL,
+        Freetype,
+        #Fluidsynth,
     ]
 
-    for build in to_build:
+    build_softwares = [b(args) for b in to_build]
+
+    for build in build_softwares:
         with Github.LogGroup(f"Building {build.name}"):
             build.build()
 
@@ -45,8 +52,8 @@ if __name__ == "__main__":
             build.publish()
 
     dump_build_notes(
-        "native-build (Linux x64)",
+        "native-build (Linux)",
         ROOT_DIR,
         ARTIFACT_DIR.joinpath("notes.md"),
-        [f"- {build.name}" for build in to_build],
+        [f"- {build.name}" for build in build_softwares],
     )
