+++ scripts/software/glfw.py
@@ -3,12 +3,13 @@
 import subprocess
 import shutil
 from common import Software, Github, Platform
-from common.cmake import cmake_common_args
+from common.cmake import cmake_common_args, locate_cmake
+from common.args import BuildArgs
 
 
 class GLFW(Software):
-    def __init__(self) -> None:
-        super().__init__("GLFW", "glfw")
+    def __init__(self, args: BuildArgs) -> None:
+        super().__init__(args, "GLFW", "glfw")
 
         self.outputs = {
             Platform.Windows: [
@@ -20,7 +21,7 @@ class GLFW(Software):
         }
 
     def build(self) -> None:
-        cmake = shutil.which("cmake")
+        cmake = locate_cmake()
 
         Github.log("Setting up CMake...")
         result = subprocess.call(
@@ -32,7 +33,7 @@ class GLFW(Software):
                 "-DGLFW_BUILD_TESTS=Off",
                 "-DGLFW_BUILD_DOCS=Off",
                 "-DGLFW_INSTALL=Off",
-            ] + cmake_common_args(),
+            ] + cmake_common_args(self.build_args),
             text=True,
             cwd=self.source_dir,
         )
